CLOUDSTACK-4459: one more try, bypass libvirt to find volume if libvirt call failed
Conflicts:

	server/src/com/cloud/template/TemplateManagerImpl.java

+++ server/src/com/cloud/template/TemplateManagerImpl.java
@@ -770,29 +770,42 @@ public class TemplateManagerImpl extends ManagerBase implements TemplateManager,
     }
 
     @Override
+    @DB
     public void evictTemplateFromStoragePool(VMTemplateStoragePoolVO templatePoolVO) {
-        StoragePool pool = (StoragePool) _dataStoreMgr.getPrimaryDataStore(templatePoolVO.getPoolId());
-        VMTemplateVO template = _tmpltDao.findByIdIncludingRemoved(templatePoolVO.getTemplateId());
-
-        if (s_logger.isDebugEnabled()) {
-            s_logger.debug("Evicting " + templatePoolVO);
+        //Need to hold the lock, otherwise, another thread may create a volume from the template at the same time.
+        //Assumption here is that, we will hold the same lock during create volume from template
+        VMTemplateStoragePoolVO templatePoolRef = _tmpltPoolDao.acquireInLockTable(templatePoolVO.getId());
+        if (templatePoolRef == null) {
+           s_logger.debug("can't aquire the lock for template pool ref:" + templatePoolVO.getId());
+           return;
         }
-        DestroyCommand cmd = new DestroyCommand(pool, templatePoolVO);
 
         try {
-            Answer answer = _storageMgr.sendToPool(pool, cmd);
+            StoragePool pool = (StoragePool) this._dataStoreMgr.getPrimaryDataStore(templatePoolVO.getPoolId());
+            VMTemplateVO template = _tmpltDao.findByIdIncludingRemoved(templatePoolVO.getTemplateId());
 
-            if (answer != null && answer.getResult()) {
-                // Remove the templatePoolVO
-                if (_tmpltPoolDao.remove(templatePoolVO.getId())) {
-                    s_logger.debug("Successfully evicted template: " + template.getName() + " from storage pool: " + pool.getName());
-                }
-            } else {
-                s_logger.info("Will retry evicte template: " + template.getName() + " from storage pool: " + pool.getName());
+            if (s_logger.isDebugEnabled()) {
+                s_logger.debug("Evicting " + templatePoolVO);
             }
-        } catch (StorageUnavailableException e) {
-            s_logger.info("Storage is unavailable currently.  Will retry evicte template: " + template.getName() + " from storage pool: "
-                    + pool.getName());
+            DestroyCommand cmd = new DestroyCommand(pool, templatePoolVO);
+
+            try {
+                Answer answer = _storageMgr.sendToPool(pool, cmd);
+
+                if (answer != null && answer.getResult()) {
+                    // Remove the templatePoolVO
+                    if (_tmpltPoolDao.remove(templatePoolVO.getId())) {
+                        s_logger.debug("Successfully evicted template: " + template.getName() + " from storage pool: " + pool.getName());
+                    }
+                } else {
+                    s_logger.info("Will retry evicte template: " + template.getName() + " from storage pool: " + pool.getName());
+                }
+            } catch (StorageUnavailableException e) {
+                s_logger.info("Storage is unavailable currently.  Will retry evicte template: " + template.getName() + " from storage pool: "
+                        + pool.getName());
+            }
+        } finally {
+            _tmpltPoolDao.releaseFromLockTable(templatePoolRef.getId());
         }
 
     }
